@@ -2,7 +2,7 @@ package com.example.digilavita
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import kotlinx.android.synthetic.main.forgot_pass.*
 
 class ForgotPassword: AppCompatActivity() {
